idsideAI_Complete_PyCharm_Ready/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
SQLAlchemy==2.0.29
aiosqlite==0.19.0
Jinja2==3.1.3
python-dotenv==1.0.1
httpx==0.27.0
PyYAML==6.0.1
pandas==2.2.2
